requirements.txt
streamlit
panda